[Issue 1352] update parsing XCM params
