+++ packages/extension-base/src/koni/api/xcm/polkadotXcm.ts
@@ -2,49 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { _ChainAsset, _ChainInfo } from '@subwallet/chain-list/types';
-import { getBeneficiary, getDestWeight } from '@subwallet/extension-base/koni/api/xcm/utils';
-import { _getSubstrateParaId, _getXcmAssetMultilocation, _isSubstrateParaChain, _isSubstrateRelayChain } from '@subwallet/extension-base/services/chain-service/utils';
+import { getBeneficiary, getDestinationChainLocation, getDestWeight, getTokenLocation } from '@subwallet/extension-base/koni/api/xcm/utils';
+import { _isSubstrateRelayChain } from '@subwallet/extension-base/services/chain-service/utils';
 
 import { ApiPromise } from '@polkadot/api';
 
-function getDestinationChainLocation (destinationChainInfo: _ChainInfo) {
-  if (_isSubstrateParaChain(destinationChainInfo)) {
-    return {
-      V1: {
-        parents: 1,
-        interior: {
-          X1: { Parachain: _getSubstrateParaId(destinationChainInfo) }
-        }
-      }
-    };
-  }
-
-  return { // to relaychain by default
-    V1: {
-      parents: 1,
-      interior: 'Here'
-    }
-  };
-}
-
-function getAssetLocation (tokenInfo: _ChainAsset, sendingValue: string) {
-  const multilocation = _getXcmAssetMultilocation(tokenInfo);
-
-  return {
-    V1: [
-      {
-        id: multilocation,
-        fun: { Fungible: sendingValue }
-      }
-    ]
-  };
-}
-
 export function getExtrinsicByPolkadotXcmPallet (tokenInfo: _ChainAsset, originChainInfo: _ChainInfo, destinationChainInfo: _ChainInfo, recipientAddress: string, value: string, api: ApiPromise) {
   const weightParam = getDestWeight();
-  const beneficiary = getBeneficiary(originChainInfo, destinationChainInfo, recipientAddress);
+  const beneficiary = getBeneficiary(destinationChainInfo, recipientAddress);
   const destination = getDestinationChainLocation(destinationChainInfo);
-  const assetLocation = getAssetLocation(tokenInfo, value);
+  const assetLocation = getTokenLocation(tokenInfo, value);
 
   let method = 'limitedReserveTransferAssets';
